gemini_api/requirements.txt
requests==2.32.3
beautifulsoup4==4.10.0
bs4==0.0.1
google-generativeai>=0.1.0
python-dateutil==2.9.0
python-dotenv==1.0.1
flask-cors==3.0.10
typing-extensions==4.1.1
flask==3.0.3
gunicorn>=20.1.0
re
nltk
numpy
gensim
networkx
scipy